added adjustments to PrivatBank
