--- harmashProj/src/main/java/api/CurrencyPrivatDTO.java
@@ -1,66 +1,0 @@
-package api;
-
-import com.google.gson.annotations.SerializedName;
-
-public class CurrencyPrivatDTO {
-    @SerializedName("ccy")
-    String ccy;
-    @SerializedName("base_ccy")
-    String base_ccy;
-    @SerializedName("buy")
-    String buy;
-    @SerializedName("sale")
-    String sale;
-
-    public CurrencyPrivatDTO(String ccy, String base_ccy, String buy, String sale) {
-        this.ccy = ccy;
-        this.base_ccy = base_ccy;
-        this.buy = buy;
-        this.sale = sale;
-    }
-
-    public CurrencyPrivatDTO() {
-    }
-
-    public String getCcy() {
-        return ccy;
-    }
-
-    public void setCcy(String ccy) {
-        this.ccy = ccy;
-    }
-
-    public String getBase_ccy() {
-        return base_ccy;
-    }
-
-    public void setBase_ccy(String base_ccy) {
-        this.base_ccy = base_ccy;
-    }
-
-    public String getBuy() {
-        return buy;
-    }
-
-    public void setBuy(String buy) {
-        this.buy = buy;
-    }
-
-    public String getSale() {
-        return sale;
-    }
-
-    public void setSale(String sale) {
-        this.sale = sale;
-    }
-
-    @Override
-    public String toString() {
-        return "CurrencyPrivatDTO{" +
-                "ccy='" + ccy + '\'' +
-                ", base_ccy='" + base_ccy + '\'' +
-                ", buy='" + buy + '\'' +
-                ", sale='" + sale + '\'' +
-                '}';
-    }
-}
