--- harmashProj/src/main/java/api/EndPointsPrivatBank.java
@@ -1,6 +1,0 @@
-package api;
-
-public interface EndPointsPrivatBank {
-    String baseUrl ="https://api.privatbank.ua";
-    String EXCHANGE_CURRENCY = baseUrl + "/p24api/pubinfo?exchange&json&coursid=11";
-}
