+++ harmashProj/src/main/java/api/PrivatBank/EndPointsPrivatBank.java
@@ -1,0 +1,6 @@
+package api.PrivatBank;
+
+public interface EndPointsPrivatBank {
+    String baseUrl ="https://api.privatbank.ua";
+    String EXCHANGE_CURRENCY = baseUrl + "/p24api/pubinfo";
+}
